+++ tetra/js/tetra.core.js
@@ -13,14 +13,14 @@ const Tetra = {
     return {
       // Alpine.js lifecycle:
       init() {
-        this.$dispatch('tetra:child-component-init', {component:  this});
+        this.$dispatch('tetra:childComponentInit', {component:  this});
         this.__initServerWatchers();
         if (this.__initInner) {
           this.__initInner();
         }
       },
       destroy() {
-        this.$dispatch('tetra:child-component-destroy', {component:  this});
+        this.$dispatch('tetra:childComponentDestroy', {component:  this});
         if (this.__destroyInner) {
           this.__destroyInner();
         }
@@ -50,13 +50,13 @@ const Tetra = {
           },
           lookahead: true
         });
-        this.$dispatch('tetra:component-updated', { component: this });
+        this.$dispatch('tetra:componentUpdated', { component: this });
       },
       _updateData(data) {
         for (const key in data) {
           this[key] = data[key];
         }
-        this.$dispatch('tetra:component-data-updated', { component: this });
+        this.$dispatch('tetra:componentDataUpdated', { component: this });
       },
       _setValueByName(name, value){
         // sets value to the input field with the given name
@@ -67,14 +67,14 @@ const Tetra = {
         }
       },
       _removeComponent() {
-        this.$dispatch('tetra:component-before-remove', { component: this });
+        this.$dispatch('tetra:componentBeforeRemove', { component: this });
         this.$root.remove();
       },
       _replaceComponent(html) {
-        this.$dispatch('tetra:component-before-remove', { component: this });
+        this.$dispatch('tetra:componentBeforeRemove', { component: this });
         this.$root.insertAdjacentHTML('afterend', html);
         this.$root.remove();
-        this.$dispatch('tetra:component-updated', { component: this });
+        this.$dispatch('tetra:componentUpdated', { component: this });
       },
       _redirect(url) {
         document.location = url;
@@ -106,7 +106,7 @@ const Tetra = {
       },
       __childComponents: {},
       __rootBind: {
-        ['@tetra:child-component-init'](event) {
+        ['@tetra:childComponentInit'](event) {
           event.stopPropagation();
           const comp = event.detail.component;
           if (comp.key === this.key) {
@@ -117,7 +117,7 @@ const Tetra = {
           }
           comp._parent = this;
         },
-        ['@tetra:child-component-destroy'](event) {
+        ['@tetra:childComponentDestroy'](event) {
           event.stopPropagation();
           const comp = event.detail.component;
           if (comp.key === this.key) {
@@ -217,11 +217,11 @@ const Tetra = {
         console.error("Response is not a Tetra response. Please check the server implementation.");
         return
       }
-      // handle Django messages and emit "tetra:newmessage" for each one, so components can react on that individually
+      // handle Django messages and emit "tetra:newMessage" for each one, so components can react on that individually
       const messages = Tetra.jsonDecode(response.headers.get('T-Messages'));
       if (messages) {
         messages.forEach((message, index) => {
-          component.$dispatch('tetra:newmessage', message)
+          component.$dispatch('tetra:newMessage', message)
         })
       }
       const respData = Tetra.jsonDecode(await response.text());
@@ -297,8 +297,9 @@ const Tetra = {
       payload.body = Tetra.jsonEncode(component_state)
       payload.headers['Content-Type'] = 'application/json'
     }
-
+    component.$dispatch('tetra:beforeRequest', {component: this});
     const response = await fetch(methodEndpoint, payload);
+    component.$dispatch('tetra:afterRequest', {component: this})
     return await this.handleServerMethodResponse(response, component);
   },